refactor: use rems
