+++ frontend/src/components/Input/services/styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import colors from '../../../colors';
-import iconSizes from '../../../iconSizes';
+import colors from '../../../styleTokens/colors';
+import iconSizes from '../../../styleTokens/iconSizes';
 
 const PADDING_LEFT = '8px';
 
@@ -25,7 +25,7 @@ export const InputWrapper = styled.div`
 
 export const Input = styled.input<{ isFullwidth?: boolean }>`
     border: 1px solid ${colors.iron};
-    border-radius: 4px;
+    border-radius: 0.25rem;
     height: 40px;
     color: ${colors.shark};
     width: ${({ isFullwidth }) => (isFullwidth ? '100%' : '300px')};
