+++ frontend/src/main.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 import ReactDOM from 'react-dom';
 import './global.css';
 import { IconContext } from 'react-icons';
-import colors from './colors';
+import colors from './styleTokens/colors';
 import App from './components/App/App';
 
 const iconContextConfig = { color: colors.boulder };
